# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Weather",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/login": {
            "post": {
                "description": "Foydalanuvchi login va parol orqali tizimga kirishi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Foydalanuvchi tizimga kirishi",
                "parameters": [
                    {
                        "description": "Login ma'lumotlari",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Muvaffaqiyatli kirish",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri so'rov formati",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server xatosi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "Foydalanuvchi ma'lumotlarini qabul qilib, yangi akkaunt yaratadi",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Ro'yxatdan o'tish",
                "parameters": [
                    {
                        "description": "Ro'yxatdan o'tish ma'lumotlari",
                        "name": "register",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Muvaffaqiyatli ro'yxatdan o'tish",
                        "schema": {
                            "$ref": "#/definitions/model.RegisterResp"
                        }
                    },
                    "400": {
                        "description": "Noto'g'ri so'rov formati",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server ichki xatosi",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.LoginReq": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.RegisterReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.RegisterResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}